# Conflicts:
#	bank-service/src/main/java/rs/raf/bank_service/bootstrap/BootstrapData.java
#	user-service-kotlin/src/test/kotlin/pack/userservicekotlin/service/VerificationRequestServiceTest.kt

+++ user-service-kotlin/src/main/kotlin/pack/userservicekotlin/utils/JwtTokenUtil.kt
@@ -55,7 +55,7 @@ class JwtTokenUtil {
         val token = authHeader.replace("Bearer ", "").trim { it <= ' ' }
 
         // Parsiramo token i vadimo userId
-        return getClaimsFromToken(token).get("userId", Long::class.java)
+        return getClaimsFromToken(token).get("userId", Int::class.java).toLong()
     }
 
     companion object {
